moved chared files
