+++ src/shared/memory.h
